+++ doc/Changelog.md
@@ -3,6 +3,8 @@ Changelog {#Changelog}
 
 # git master
 
+* [121](https://github.com/BlueBrain/Brion/pull/121):
+  Performance optimizations for map and binary compartment reports
 * [119](https://github.com/BlueBrain/Brion/pull/119):
   Optimized brain::Circuit::loadMorphologies for slow stat'ing
   filesystems (GPFS)